+++ code/system.ui/react.dev/src/test.ui/entry.Specs.mts
@@ -4,5 +4,5 @@ export const ModuleSpecs = {
 
 export const SampleSpecs = {
   'sample.MySample': () => import('../test.ui/sample.specs/MySample.SPEC'),
-  'sample.DevTools.Button': () => import('../test.ui/sample.DevTools/s.ui.Button.SPEC'),
+  'sample.DevTools.Button': () => import('../test.ui/sample.DevTools/ui.Button.SPEC'),
 };
