+++ code/system.ui/react.dev/src/test.ui/sample.specs/MySample.SPEC.tsx
@@ -1,6 +1,6 @@
 import { t, Color, css, Spec } from '../common';
 import { MySample } from './MySample';
-import { SampleDevTools as DevTools } from '../sample.DevTools';
+import { DevTools } from '../sample.DevTools';
 import { DevBus } from '../../logic.Bus';
 
 let _renderCount = 0;
@@ -83,7 +83,7 @@ export default Spec.describe('MySample', (e) => {
   );
 
   e.it('SampleDevTools', async (e) => {
-    const dev = DevTools(e);
+    const dev = DevTools.curry(e);
 
     const ctx = await dev.button((e) => {
       console.log('Inside Spec - BUTTON', e);
